requirements.txt
# ── Core ────────────────────────────────────────────────
pyyaml>=6.0
openpyxl>=3.1.0
streamlit>=1.30.0

# ── Database (HITL feedback) ────────────────────────────
psycopg2-binary>=2.9.0

# ── LLM providers ──────────────────────────────────────
anthropic>=0.20.0
openai>=1.0.0

# ── Testing ─────────────────────────────────────────────
pytest>=7.0.0
pytest-cov>=4.0.0

# ── Utilities ───────────────────────────────────────────
pandas>=2.0.0
